+++ Sources/SwiftKit/Coding/StringCoder.swift
@@ -1,0 +1,15 @@
+//
+//  StringDecoder.swift
+//  SwiftKit
+//
+//  Created by Daniel Saidi on 2015-03-21.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This protocol can be implemented by classes that can encode
+ and decode strings.
+ */
+public protocol StringCoder: StringEncoder, StringDecoder {}