+++ Release Notes.md
@@ -5,10 +5,14 @@
 
 This release adjusts [Authentication][Authentication] service signatures.
 
+It also adds base64 string [coding][Coding] functionality to `SwiftKit`. 
+
 
 ## 0.0.1
 
 This release adds [Authentication][Authentication] functionality to `SwiftKit`.
 
 
-[Authentication]: Readmes/Authentication.md 
+[Authentication]: Readmes/Authentication.md
+[Coding]: Readmes/Coding.md 
+[Extensions]: Readmes/Extensions.md