+++ Release Notes.md
@@ -3,16 +3,17 @@
 
 ## 0.0.2
 
-This release adjusts [Authentication][Authentication] service signatures.
+This release adjusts [authentication][Authentication] service signatures.
 
-It also adds base64 string [coding][Coding] functionality to `SwiftKit`. 
+It also adds [coding][Coding] and [IoC][IoC] functionality to `SwiftKit`. 
 
 
 ## 0.0.1
 
-This release adds [Authentication][Authentication] functionality to `SwiftKit`.
+This release adds [authentication][Authentication] functionality to `SwiftKit`.
 
 
 [Authentication]: Readmes/Authentication.md
 [Coding]: Readmes/Coding.md 
 [Extensions]: Readmes/Extensions.md
+[IoC]: Readmes/IoC.md